# Performance test for different types of Pixel structures