--- a/.gitignore
+++ b/.gitignore
@@ -8,6 +8,7 @@ tools/
 
 # Test results
 test_results/
+BenchmarkDotNet.Artifacts/
 
 # Documentation output
 docs/_site