--- a/src/Directory.Packages.props
+++ b/src/Directory.Packages.props
@@ -4,6 +4,8 @@
         <PackageVersion Include="Yarhl" Version="3.0.0-alpha06" />
         <PackageVersion Include="System.Drawing.Common" Version="5.0.2" />
 
+        <PackageVersion Include="BenchmarkDotNet" Version="0.12.1" />
+
         <PackageVersion Include="Microsoft.SourceLink.GitHub" Version="1.0.0" />
 
         <PackageVersion Include="StyleCop.Analyzers" Version="1.1.118" />