### Description

Perform a performance test to evaluate different implementations of the elements (Pixel) of an image. The options were:

- `uint` that represents a 24-bits color or the color table index + 8-bits for alpha
  - Size: 4 bytes
- `Pixel` struct that separates the alpha channel and has a boolean to indicate if it's indexed or not.
  - Size: 8 bytes
- `PixelRgb` with a field per color channel and alpha.
  - Size: 6 bytes

``` ini

BenchmarkDotNet=v0.12.1, OS=ubuntu 20.04 (container)
Intel Core i7-4720HQ CPU 2.60GHz (Haswell), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=5.0.201
  [Host]     : .NET Core 5.0.4 (CoreCLR 5.0.421.11614, CoreFX 5.0.421.11614), X64 RyuJIT
  DefaultJob : .NET Core 5.0.4 (CoreCLR 5.0.421.11614, CoreFX 5.0.421.11614), X64 RyuJIT


```
|              Method |    Size |         Mean |      Error |     StdDev |    Gen 0 |    Gen 1 |    Gen 2 |   Allocated |
|-------------------- |-------- |-------------:|-----------:|-----------:|---------:|---------:|---------:|------------:|
| SetPrimitiveIndexed |   49152 |    124.37 μs |   0.632 μs |   0.560 μs |  62.2559 |  62.2559 |  62.2559 |   192.04 KB |
|     SetPixelIndexed |   49152 |    199.56 μs |   0.932 μs |   0.826 μs | 124.7559 | 124.7559 | 124.7559 |   384.06 KB |
|  SetPixelRgbIndexed |   49152 |    461.56 μs |   1.033 μs |   0.916 μs |  90.8203 |  90.8203 |  90.8203 |   288.05 KB |
|   SetPrimitiveColor |   49152 |    169.70 μs |   0.536 μs |   0.475 μs |  62.2559 |  62.2559 |  62.2559 |   192.04 KB |
|       SetPixelColor |   49152 |    259.35 μs |   1.615 μs |   1.432 μs | 124.5117 | 124.5117 | 124.5117 |   384.06 KB |
|    SetPixelRgbColor |   49152 |    505.18 μs |   1.663 μs |   1.474 μs |  90.8203 |  90.8203 |  90.8203 |   288.05 KB |
|   GetPrimitiveIndex |   49152 |     68.94 μs |   0.500 μs |   0.418 μs |  62.3779 |  62.3779 |  62.3779 |   192.04 KB |
|       GetPixelIndex |   49152 |    129.67 μs |   0.833 μs |   0.779 μs | 124.7559 | 124.7559 | 124.7559 |   384.06 KB |
|    GetPixelRgbIndex |   49152 |    115.20 μs |   0.787 μs |   0.698 μs |  90.8203 |  90.8203 |  90.8203 |   288.05 KB |
|   GetPrimitiveColor |   49152 |    137.52 μs |   0.594 μs |   0.555 μs |  62.2559 |  62.2559 |  62.2559 |   192.04 KB |
|       GetPixelColor |   49152 |    193.40 μs |   1.390 μs |   1.233 μs | 124.7559 | 124.7559 | 124.7559 |   384.06 KB |
|    GetPixelRgbColor |   49152 |    173.60 μs |   1.320 μs |   1.170 μs |  90.8203 |  90.8203 |  90.8203 |   288.05 KB |
| SetPrimitiveIndexed | 8847360 | 22,212.99 μs | 404.346 μs | 378.225 μs | 968.7500 | 968.7500 | 968.7500 |  34560.3 KB |
|     SetPixelIndexed | 8847360 | 36,111.93 μs | 456.344 μs | 404.537 μs | 928.5714 | 928.5714 | 928.5714 |  69120.3 KB |
|  SetPixelRgbIndexed | 8847360 | 82,676.19 μs | 632.159 μs | 591.322 μs | 833.3333 | 833.3333 | 833.3333 |    51841 KB |
|   SetPrimitiveColor | 8847360 | 30,725.33 μs | 450.606 μs | 399.451 μs | 968.7500 | 968.7500 | 968.7500 |  34560.3 KB |
|       SetPixelColor | 8847360 | 47,022.38 μs | 486.220 μs | 406.016 μs | 909.0909 | 909.0909 | 909.0909 | 69121.11 KB |
|    SetPixelRgbColor | 8847360 | 90,049.44 μs | 442.200 μs | 413.634 μs | 833.3333 | 833.3333 | 833.3333 |    51841 KB |
|   GetPrimitiveIndex | 8847360 | 10,812.43 μs | 107.532 μs |  95.324 μs | 984.3750 | 984.3750 | 984.3750 |  34560.3 KB |
|       GetPixelIndex | 8847360 | 20,093.64 μs | 166.288 μs | 155.546 μs | 968.7500 | 968.7500 | 968.7500 |  69120.3 KB |
|    GetPixelRgbIndex | 8847360 | 18,418.86 μs | 103.715 μs |  91.941 μs | 968.7500 | 968.7500 | 968.7500 |  51840.3 KB |
|   GetPrimitiveColor | 8847360 | 23,057.97 μs | 129.346 μs | 114.662 μs | 968.7500 | 968.7500 | 968.7500 |  34560.3 KB |
|       GetPixelColor | 8847360 | 31,531.98 μs | 189.125 μs | 176.908 μs | 937.5000 | 937.5000 | 937.5000 |  69120.3 KB |
|    GetPixelRgbColor | 8847360 | 28,801.54 μs | 137.747 μs | 115.025 μs | 968.7500 | 968.7500 | 968.7500 |  51840.3 KB |


## Conclusion

Even with an 4k image it takes just 28 ms to read and 47 ms to write with the structure `PixelRgb`. As that structure is easier to work with and the time is quite fast it will be the option to go. The size of an image would be 6 times bytes the number of pixels. For a 4k image it would be 50 MB. For a NDS image it would be 288 KB. By reusing buffers this would be acceptable.